backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
python-dotenv>=1.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.7
bcrypt>=4.1.2
email-validator>=2.1.0.post1
pydantic>=2.6.0
pydantic-settings>=2.1.0
cryptography>=42.0.0 